+++ src/main/java/com/grow/payment_service/payment/presentation/controller/PaymentQueryController.java
@@ -1,4 +1,4 @@
-package com.grow.payment_service.payment.presentation;
+package com.grow.payment_service.payment.presentation.controller;
 
 import java.util.List;
