+++ src/main/java/com/grow/payment_service/payment/presentation/controller/WebhookController.java
@@ -1,6 +1,5 @@
-package com.grow.payment_service.payment.presentation;
+package com.grow.payment_service.payment.presentation.controller;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
